backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
langchain-groq==0.0.1
chromadb==0.4.22
openai==1.10.0
groq==0.4.2
python-dotenv==1.0.0
pypdf==4.0.1
python-docx==1.1.0
duckduckgo-search==4.4.3
tiktoken==0.5.2
pydantic==2.6.0
sse-starlette==2.0.0
httpx==0.26.0